% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISN.R
\name{BISN}
\alias{BISN}
\title{Bayesian Inference of Sparse Networks}
\usage{
BISN(data, eta = 300, max_iter = 10000, tol = 0.01, r = 0.5,
  s = ncol(data)/(0.001 * (ncol(data) - 1) + 1), backward_pass = TRUE,
  prm_learning = FALSE)
}
\arguments{
\item{data}{n x p matrix of observed data. n is the sample size and p is the dimension. Missing data can be denoted by NA.}

\item{eta}{upper bound of the step size (eta >0, 300 by default).}

\item{max_iter}{maximum number of iterations (1e4 by default).}

\item{tol}{tolerance to check convergence of the algorithm (1e-2 by default).}

\item{r}{the decaying factor (0 < r < 1, 0.5 by default)}

\item{s}{minibatch size (s = p/(1e-3*(p-1)+1)) by default)}

\item{backward_pass}{boolean value to decide whether to enable the backward pass or not (1 by default). 
backward_pass = 1 would run the BISN algorithm again by reversely ordering the data (i.e., the backward pass), 
and then average the results from both forward and backward pass. This will improve the estimation accuracy, 
especially when the sample size is small.}

\item{prm_learning}{boolean value to decide whether to reestimate the non-zero elements via maximum 
likelihood or not. (0 by default)}
}
\value{
Ksparse p x p matrix with the same zero pattern as the estimated adjacency matrix Adj. 
The nonzero elements in Ksparse is reestimated by maximum likelihood if prm_learning = TRUE.

Adj estimated adjacency matrix using the method in Section V in (Yu et al, Variational wishart approximation for 
graphical model selection: Monoscale and multiscale models, 2019).

Kest p x p full matrix. Kest = ML * MD * ML', where MD and ML denotes the mean of the D and L matrix.

Lambda p x p estimated Lambda matrix.

run_time total running time.
}
\description{
Bayesian Inference of Sparse Networks
}
